Move global celery hooks to corehq.celery_monitoring.signals
